Fix WEBAPP_FOLDER cleaning in test removing the project root folder

+++ flow-server/src/test/java/com/vaadin/flow/server/DevModeHandlerTest.java
@@ -70,6 +70,7 @@ public class DevModeHandlerTest {
     private static final String TEST_FILE = "webpack-out.test";
     private HttpServer httpServer;
     private int responseStatus;
+    private boolean webpackFolderCreated = false;
 
     @Rule
     public ExpectedException exception = ExpectedException.none();
@@ -82,7 +83,11 @@ public class DevModeHandlerTest {
     }
 
     private void createWebpackScript(String readyString, int milliSecondsToRun) throws IOException {
-        new File(WEBAPP_FOLDER).mkdirs();
+        File dir = new File(WEBAPP_FOLDER);
+        webpackFolderCreated = !dir.isDirectory();
+        if (webpackFolderCreated) {
+          new File(WEBAPP_FOLDER).mkdirs();
+        }
         File serverFile = new File(WEBPACK_SERVER);
         serverFile.getParentFile().mkdirs();
         serverFile.createNewFile();
@@ -100,7 +105,9 @@ public class DevModeHandlerTest {
     @After
     public void teardown() throws IOException {
         FileUtils.deleteDirectory(new File("node_modules"));
-        FileUtils.deleteDirectory(new File(WEBAPP_FOLDER));
+        if (webpackFolderCreated) {
+          FileUtils.deleteDirectory(new File(WEBAPP_FOLDER));
+        }
         FileUtils.deleteQuietly(new File(WEBPACK_CONFIG));
         if (httpServer != null) {
             httpServer.stop(0);
@@ -164,10 +171,12 @@ public class DevModeHandlerTest {
 
     @Test
     public void shouldNot_CreateInstance_When_WebappFolderNotFound() throws Exception {
-        Thread.sleep(150);
-        FileUtils.deleteDirectory(new File(WEBAPP_FOLDER));
-        assertNull(createInstance(configuration));
-        Thread.sleep(150);
+        if (webpackFolderCreated) {
+          Thread.sleep(150);
+          FileUtils.deleteDirectory(new File(WEBAPP_FOLDER));
+          assertNull(createInstance(configuration));
+          Thread.sleep(150);
+        }
     }
 
     @Test
